Avoid creating multiple tasks for config entry init
I added a new task in #49241 to ensure we could cancel the discovery flow at shutdown.  Instead of making a new task we can use the interrupt context manager to cancel the awaited coro when a future resolves with `async-interrupt`.  `async-interrupt` is based on the original design of `async-timeout`. It’s currently used in aioesphomeapi and a few other places.  It was originally built to solve a race problem with esphome Bluetooth proxies but no longer used directly after #104644